Rename all `lua5{1,2,3,4}` features to `lua5_{1,2,3,4}`
