+++ src/rust/iced-x86-lua/Cargo.toml
@@ -10,10 +10,10 @@ crate-type = ["cdylib"]
 
 [features]
 default = []
-lua51 = ["loona/lua51"]
-lua52 = ["loona/lua52"]
-lua53 = ["loona/lua53"]
-lua54 = ["loona/lua54"]
+lua5_1 = ["loona/lua5_1"]
+lua5_2 = ["loona/lua5_2"]
+lua5_3 = ["loona/lua5_3"]
+lua5_4 = ["loona/lua5_4"]
 extra_checks = ["loona/extra_checks"]
 
 [dependencies]
