+++ src/rust/loona/Cargo.toml
@@ -8,10 +8,10 @@ publish = false
 
 [features]
 default = []
-lua51 = []
-lua52 = []
-lua53 = []
-lua54 = []
+lua5_1 = []
+lua5_2 = []
+lua5_3 = []
+lua5_4 = []
 extra_checks = []
 
 [dependencies]
