+++ src/rust/loona/src/lua_api.rs
@@ -3,20 +3,20 @@
 
 //! Low level access to the Lua C API
 
-#[cfg(feature = "lua51")]
-mod lua51;
-#[cfg(feature = "lua52")]
-mod lua52;
-#[cfg(feature = "lua53")]
-mod lua53;
-#[cfg(feature = "lua54")]
-mod lua54;
+#[cfg(feature = "lua5_1")]
+mod lua5_1;
+#[cfg(feature = "lua5_2")]
+mod lua5_2;
+#[cfg(feature = "lua5_3")]
+mod lua5_3;
+#[cfg(feature = "lua5_4")]
+mod lua5_4;
 
-#[cfg(feature = "lua51")]
-pub use crate::lua_api::lua51::*;
-#[cfg(feature = "lua52")]
-pub use crate::lua_api::lua52::*;
-#[cfg(feature = "lua53")]
-pub use crate::lua_api::lua53::*;
-#[cfg(feature = "lua54")]
-pub use crate::lua_api::lua54::*;
+#[cfg(feature = "lua5_1")]
+pub use crate::lua_api::lua5_1::*;
+#[cfg(feature = "lua5_2")]
+pub use crate::lua_api::lua5_2::*;
+#[cfg(feature = "lua5_3")]
+pub use crate::lua_api::lua5_3::*;
+#[cfg(feature = "lua5_4")]
+pub use crate::lua_api::lua5_4::*;
